CLEANSalesAgentProject/requirements.txt
# Core Flask dependencies
Flask==2.2.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2

# API and authentication
anthropic==0.6.0
authlib==1.2.0
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.7

# Security
cryptography==40.0.1
pyOpenSSL==23.1.1

# Utilities
gunicorn==20.1.0
pytz==2023.3

# Testing (optional)
pytest==7.3.1